Update for second submission

+++ application.py
@@ -39,7 +39,7 @@ def showRestaurants():
 
 # Create a new restaurant
 @app.route('/restaurants/new/', methods=['GET', 'POST'])
-def newRestaurant(restaurant_id):
+def newRestaurant():
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
@@ -61,6 +61,9 @@ def editRestaurant(restaurant_id):
         id=restaurant_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+    if editedRestaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()'>"
+    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         if request.form['name']:
             editedRestaurant.name = request.form['name']
@@ -79,6 +82,9 @@ def deleteRestaurant(restaurant_id):
         id=restaurant_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()'>"
+    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         session.delete(restaurantToDelete)
         flash("Successfully Deleted %s" % restaurantToDelete.name)
@@ -98,12 +104,13 @@ def showMenu(restaurant_id):
         id=restaurant_id).one()
     items = session.query(MenuItem).filter_by(
         restaurant_id=restaurant_id).all()
-    if 'username' not in login_session:
+    creator = getUserInfo(restaurant.user_id)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
         return render_template('publicMenu.html', items=items,
-                               restaurant=restaurant)
+                               restaurant=restaurant, creator=creator)
     else:
         return render_template('menu.html', restaurant=restaurant,
-                               items=items)
+                               items=items, creator=creator)
 
 
 # Create a new menu item
@@ -113,16 +120,18 @@ def newMenuItem(restaurant_id):
     if 'username' not in login_session:
         return redirect('/login')
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
-    if request.method == 'POST':
-        newItem = MenuItem(name=request.form['name'],
-                           description=request.form['description'],
-                           price=request.form['price'],
-                           course=request.form['course'],
-                           restaurant_id=restaurant_id)
-        session.add(newItem)
-        flash("Menu Item '%s' Successfully Created!" % newItem.name)
-        session.commit()
-        return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()'>"
+        if request.method == 'POST':
+            newItem = MenuItem(name=request.form['name'],
+                               description=request.form['description'],
+                               price=request.form['price'],
+                               course=request.form['course'],
+                               restaurant_id=restaurant_id)
+            session.add(newItem)
+            flash("Menu Item '%s' Successfully Created!" % newItem.name)
+            session.commit()
+            return redirect(url_for('showMenu', restaurant_id=restaurant_id))
     else:
         return render_template('newMenuItem.html', restaurant_id=restaurant_id)
 
@@ -133,8 +142,10 @@ def newMenuItem(restaurant_id):
 def editMenuItem(restaurant_id, menu_id):
     if 'username' not in login_session:
         return redirect('/login')
-    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -162,6 +173,8 @@ def deleteMenuItem(restaurant_id, menu_id):
         return redirect('/login')
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(itemToDelete)
         flash("Menu Item '%s' Successfully Deleted" % itemToDelete.name)
